Move init.c to internal.c
